+++ events/reviews.go
@@ -89,6 +89,11 @@ func processReview(review models.Review) {
 			detection.TopScore = detection.Data.TopScore
 		}
 
+		// Wait for license plate data before notifying, if set
+		if config.ConfigData.Alerts.LicensePlate.Enabled {
+			waitforLPR(&detection)
+		}
+
 		// Check that event passes configured filters
 		detection.CurrentZones = detection.Zones
 		if !checkEventFilters(detection) {
@@ -142,6 +147,7 @@ func recheckReview(review models.Review) models.Review {
 		log.Error().
 			Err(err).
 			Msgf("Cannot get event from %s", url)
+		return review
 	}
 	config.Internal.Status.Health = "ok"
 	config.Internal.Status.Frigate.API = "ok"