+++ models/internal.go
@@ -29,13 +29,14 @@ type Notifiers struct {
 	AppriseAPI []NotifierStatus `json:"apprise-api" doc:"Status of Apprise API notifications"`
 	Discord    []NotifierStatus `json:"discord" doc:"Status of Discord notifications"`
 	Gotify     []NotifierStatus `json:"gotify" doc:"Status of Gotify notifications"`
+	Matrix     []NotifierStatus `json:"matrix" doc:"Status of Matrix notifications"`
+	Mattermost []NotifierStatus `json:"mattermost" doc:"Status of Mattermost notifications"`
 	Ntfy       []NotifierStatus `json:"ntfy" doc:"Status of Ntfy notifications"`
 	Pushover   []NotifierStatus `json:"pushover" doc:"Status of Pushover notifications"`
 	Signal     []NotifierStatus `json:"signal" doc:"Status of Signal notifications"`
 	SMTP       []NotifierStatus `json:"smtp" doc:"Status of SMTP notifications"`
 	Telegram   []NotifierStatus `json:"telegram" doc:"Status of Telegram notifications"`
 	Webhook    []NotifierStatus `json:"webhook" doc:"Status of Webhook notifications"`
-	Mattermost []NotifierStatus `json:"mattermost" doc:"Status of Mattermost notifications"`
 }
 
 type NotifierStatus struct {