+++ docs/changelog.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [v0.4.4](https://github.com/0x2142/frigate-notify/releases/tag/v0.4.4) - Apr 08 2025
+ - Add support for notifications via [Matrix](https://frigate-notify.0x2142.com/latest/config/file/#matrix)
+ - Add support for `message_thread_id` to [Telegram](https://frigate-notify.0x2142.com/latest/config/file/#telegram) notifications
+ - Add review/event link for Pushover action button
+ - Add ability to disable Discord notifications being sent as embedded message
+ - Add ability to wait for [license plate recognition](https://frigate-notify.0x2142.com/latest/config/file/#license-plate) data from Frigate to be included in notifications
+ - Add ability to allow / block license plates from generating notification
+ - Fix issue with Webhook default message template
+
 ## [v0.4.3](https://github.com/0x2142/frigate-notify/releases/tag/v0.4.3) - Feb 21 2025
  - Add support for notifications via [Apprise API](https://frigate-notify.0x2142.com/latest/config/file/#apprise-api)
  - Add retry interval for collecting snapshots from Frigate